app/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
psycopg2-binary
SQLAlchemy==2.0.38
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1